Refactored dropzone and reset button
